requirements.txt
requests
pandas
beautifulsoup4
xlrd
pyarrow
pyspark
apache-airflow[postgres]
apache-airflow-providers-apache-spark
geopy
